Replace Whoosh search by PostgreSQL full-text search
Whoosh isn’t actively maintained. It’s an external component that slows
down deployment because it needs a local index generated. It also
hinders horizontal scaling as all workers would need a replica of the
index file.

Ideally, search would have been either in public forums with a
description (maintained by the communaute staff), or in a topic where
a single result would show for all of its posts.
Whoosh did not offer customization to index all posts of a topic
together.

Instead, generate an index for search in PostgreSQL. That allows a lot
of flexibility, and the database is always available and shared with all
workers.

The index is cached using a materialized view, that is refreshed
periodically. The refresh process takes less than 200 ms currently, so
an update every 10 minutes isn’t likely to have noticeable performance
implications.

To query the index and present search results, the search django app was
created. It contains a model `managed=False` to access the materialized
view, and the wiring to expose the search capability to users.

Some small differences arose when porting the templates:
- with posts aggregated under a single topic, the poster username and
  post creation date were no longer available. They aren’t essential to
  search results, simply dropped them.
- `query` was no longer available in context, use the form.q.value
  instead to retrieve the search terms.
- `page` was defined by haystack, Django ListView provides `page_obj`
  instead.

Because databases are isolated in tests by `pytest_django`, the search
index also gain isolation without extra work, resolving some
long-standing test concurrency issues.