Remove storage from PublicFile migrations changes
The storage name changes with each environment, causing migration checks
to fail because the model has changes not reflected in a migration.
Instead, use the current environment storage name in the migration so
that it always matches. This value is of little interest for migrations
anyways, as it does not relate to the database.